+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add new early game weapon: Wizard Staff
 - Add new texture for Novice Wand
+- Add emissive textures for weapons
 - Frostbolt slowness status effect emits snowflake particles
 - Rebalance Arcane Staff, Fire Staff, Frost Staff durability and enchantability (now matching diamond tier)
 - Fix luminance of spell projectiles and effects