+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.9.23
+
+- Add quartz to Villager Wizard chest loot table
+- Update Frost Staff textures
+
 # 0.9.22
 
 - Add new early game weapon: Wizard Staff